Refactor: entity 패키지명을 domain으로 수정 - #6 - 통일성을 위함
