+++ src/main/java/sopt/sopt36goodoc/qna/dto/response/QnAPreviewResponses.java
@@ -2,7 +2,7 @@ package sopt.sopt36goodoc.qna.dto.response;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
-import sopt.sopt36goodoc.qna.entity.QnA;
+import sopt.sopt36goodoc.qna.domain.QnA;
 
 import java.util.stream.Collectors;
