+++ src/main/java/sopt/sopt36goodoc/qna/repository/QnARepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import sopt.sopt36goodoc.hospital.domain.Department;
-import sopt.sopt36goodoc.qna.entity.QnA;
+import sopt.sopt36goodoc.qna.domain.QnA;
 
 import java.util.List;
